+++ src/app/pages/demo-angular-lib/demo-angular-lib-routing.module.ts
@@ -1,13 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DemoComponent } from './demo/demo.component';
-import { DemoPrismjsHighlightComponent } from './demo-components/demo-prismjs-highlight/demo-prismjs-highlight.component';
 
 
 const routes: Routes = [
   { path: '', redirectTo: 'demo', pathMatch: 'full' },
   {
-    path: 'modal-component', loadChildren: () => import('./demo-components/demo-modal-component/demo-modal-component.module').then(m => m.DemoModalComponentModule)
+    path: 'modal-component',
+    loadComponent: () => import('./demo-components/demo-modal-component/demo-modal-component.component').then(c => c.DemoModalComponent)
   },
   {
     path: 'form-components',
@@ -23,12 +22,12 @@ const routes: Routes = [
   },
   {
     path: 'prismjs-highlight',
-    component: DemoPrismjsHighlightComponent
+    loadComponent: () => import('./demo-components/demo-prismjs-highlight/demo-prismjs-highlight.component').then(c => c.DemoPrismjsHighlightComponent)
   },
   { path: 'pipes', loadChildren: () => import('./demo-pipes/demo-pipes.module').then(m => m.DemoPipesModule) },
   {
     path: 'demo',
-    component: DemoComponent
+    loadComponent: () => import('./demo/demo.component').then(c => c.DemoComponent)
   },
 ];