+++ src/app/pages/demo-angular-lib/demo-components/demo-filter-components/intro/demo-filter-components.component.ts
@@ -1,8 +1,8 @@
 import { JsonPipe } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { SampleProductsData } from 'src/app/shared/constant/products';
-import { MdsHightlightPrismModule, MdsFilterModule, SelectedFilterSliderInterface }
+import { ProductDataModel, SampleProductsData } from 'src/app/shared/constant/products';
+import { MdsHightlightPrismModule, MdsFilterModule, MdsFilterModel }
 /*-public-mode-*/ from 'medes-ui';
 //*-dev-mode-*/ from 'projects/medes-ui/src/public-api';
 
@@ -14,67 +14,63 @@ import { MdsHightlightPrismModule, MdsFilterModule, SelectedFilterSliderInterfac
   imports: [RouterModule, JsonPipe, MdsHightlightPrismModule, MdsFilterModule]
 })
 export class DemoFilterComponentsComponent implements OnInit {
-  sampledata = [];
-  selected = {};
-  colormap = {
-      'blue-sky': '#66ccdd',
-      maroon: '#bb6a66',
-      brown: '#6b6a6b',
-      black: '#000000',
-      purple: '#6a66bb',
-      gold: '#b6ab66'
-  };
-  filterRangeSelected: SelectedFilterSliderInterface = {
-    min: 20,
-    max: 120,
-    start: 25,
-    end: 50
-  };
+  sampledata: ProductDataModel[];
+  mdsFilterModel: MdsFilterModel;
 
 checkboxComponent = `
-<!-- Filter Checkboxes Component - Multiple Group -->
-<mds-filter-checkbox
-  [titlegroup]="'Filter by Category'"
-  [filterData]="sampledata"
-  [filterBy]="'category'"
-  [filterSelected]="selected"
-  [reset]="'✕'"
-  [hideCounter]="false"
-></mds-filter-checkbox>`;
+<!-- Filter Checkboxes Component -->
+<mds-filter-checkbox id="by-category" [data]="data" [model]="mdsFilterModel"></mds-filter-checkbox>`;
 
 swatchComponent = `
 <!-- Filter Swatch Component -->
-<mds-filter-swatch
-  [titlegroup]="'Filter by Color'"
-  [filterData]="sampledata"
-  [filterBy]="'color'"
-  [filterSelected]="selected"
-  [swatchMapping]="colormap"
-  [swatchSize]="30"
-  [swatchRadius]="25"
-></mds-filter-swatch>`;
+<mds-filter-swatch id="color" [data]="data" [model]="mdsFilterModel"></mds-filter-swatch>`;
 
 rangeSliderComponent = `
 <!-- Demo Filter Range Slider Component -->
-<mds-filter-range-slider
-  [label]="'Filter by something'"
-  [min]="20"
-  [max]="100"
-  [(start)]="25"
-  [(end)]="40"
-  [disabled]="false"
-></mds-filter-range-slider>`;
+<mds-filter-range-slider [model]="mdsFilterModel"></mds-filter-range-slider>`;
 
 resetComponent = `
 <!-- Reset Filter Component -->
-<mds-filter-reset
-  [content]="'Reset All'"
-  [(filterSelected)]="selected"
-  [(filterRangeSelected)]="filterRangeSelected">
-</mds-filter-reset>`;
+<mds-filter-reset customClass="btn-swatch-a" [(model)]="mdsFilterModel">Reset All</mds-filter-reset>`;
 
   ngOnInit(): void {
+    // Sample Data
     this.sampledata = SampleProductsData.data;
+    // Filter Model
+    this.mdsFilterModel = {
+      configs: {
+        checkBox: {
+          'by-category' : {
+            property: 'category',
+            label: 'Filter by Categories'
+          },
+        },
+        swatchBox: {
+          color: {
+            property: 'color',
+            label: 'Filter by Color',
+            colorMap: {
+              'blue-sky': '#66ccdd',
+              'maroon': '#bb6a66',
+              'brown': '#6b6a6b',
+              'black': '#000000',
+              'purple': '#6a66bb',
+              'gold': '#b6ab66'
+            },
+            swatchSize: 30,
+            swatchRadius: 5
+          }
+        },
+        rangeSlider: {
+          price: {
+            label: 'Filter by Price',
+            min: 0,
+            max: 2500,
+          }
+        }
+      }
+    }
+    
   }
 
 }