+++ src/app/pages/demo-angular-lib/demo-components/demo-navigation-manager/demo-page-navigation/page-navigation-preview-code.ts
@@ -1,15 +1,17 @@
 export const htmlCode = `
 <!-- Medes Page Navigation Component -->
-<mds-page-navigation [data]="data" [(pageNavConfig)]="pageNavConfig"></mds-page-navigation>`;
+<mds-page-navigation [data]="data" [model]="pageNavConfig"></mds-page-navigation>`;
 
 
 export const interfaceCode = `
-pageNavConfig: PageNavigationConfigInterface = {
-  type: MdsPagerType | string; // Required, Pager style/type will use normal or pagination
-  itemToShow?: number; // Optional (default value is 10). Total item to show
-}
-    
-MdsPagerType {
-  STANDARD = 'standard',
-  NUMBERING = 'numbering',
+interface PageNavigationManagerModel {
+    configs: {
+        itemToShow?: number;
+        pageCounter?: {...};
+        pageNavigation?: {
+            type?: 'standard' | 'numbering';
+            color?: string;
+        }
+    },
+    data: {...};
 }`;