+++ src/app/pages/demo-ui-elements/demo-ui-elements-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { DemoGridSystemComponent } from './demo-grid-system/demo-grid-system.component';
-import { DemoImageLoaderComponent } from './demo-image-loader/demo-image-loader.component';
+import { DemoLoaderComponent } from './demo-loader/demo-loader.component';
 import { DemoBasicElementsComponent } from './demo-basic-elements/demo-basic-elements.component';
 import { DemoLayoutUtilsComponent } from './demo-layout-utils/demo-layout-utils.component';
 
@@ -10,7 +10,7 @@ const routes: Routes = [
   { path: 'basic-elements', component: DemoBasicElementsComponent },
   { path: 'layout-utils', component: DemoLayoutUtilsComponent },
   { path: 'grid-system', component: DemoGridSystemComponent },
-  { path: 'image-loader', component: DemoImageLoaderComponent }
+  { path: 'loader', component: DemoLoaderComponent }
 ];
 
 @NgModule({