+++ src/app/shared/utils/code-preview-generator.ts
@@ -1,34 +1,18 @@
-
-export function constructImportModuleCode(modules: string): string {
-return `
-import { NgModule } from '@angular/core';
-import { ${modules} } from 'medes-ui';
-
-@NgModule({
-    declarations: [ ... ],
-    imports: [
-        ${modules}
-    ]
-})
-}
-
-export class MyModule { }`;
-}
-
-export function constructComponentCode(importMdsUi: string, valuesComponent: string): string {
+export function constructComponentCode(importMdsUi: string, imports?: string, providers?: string, valuesComponent?: string): string {
 return `
 import { Component } from '@angular/core';
 import { ${importMdsUi} } from 'medes-ui';
 
 @Component({
-  selector: 'mds-my-component',
-  templateUrl: './mds-my-component.component.html',
-  styleUrls: ['./mds-my-component.component.scss']
+  selector: 'mds-app',
+  templateUrl: './mds-app.component.html',
+  styleUrls: ['./mds-app.component.scss'],
+  standalone: true,
+  imports: [${imports}]${providers ? `,
+  providers: [${providers}]`: ''}
 })
 
-export class MyComponent {
-  ...
+export class MdsAppComponent {
   ${valuesComponent}
-  ...
 }`;
 }